Work on header from Frances' mockup
