+++ src/components/AuthNavRight.js
@@ -1,10 +1,10 @@
 import React from 'react'
-import { FaDiscord, FaGithub, FaTwitter } from 'react-icons/fa'
 import { Button, Nav, NavDropdown } from 'react-bootstrap'
 import axios from 'axios'
 import config from './../config'
 import ErrorHandler from './ErrorHandler'
 import { Link } from 'react-router-dom'
+import MainSearchBar from './MainSearchBar'
 
 const handleOnClick = () => {
   axios.get(config.api.getUriPrefix() + '/logout')
@@ -16,9 +16,7 @@ const handleOnClick = () => {
 
 const AuthNavRight = () =>
   <Nav className='ml-auto metriq-navbar'>
-    <Nav.Link href='https://twitter.com/MetriqInfo' eventKey='10'> <FaTwitter color={`${'black'}`} size={25} /> </Nav.Link>
-    <Nav.Link href='http://discord.unitary.fund' eventKey='11'> <FaDiscord color={`${'black'}`} size={25} /> </Nav.Link>
-    <Nav.Link href='https://github.com/unitaryfund/metriq-app' eventKey='12'> <FaGithub color={`${'black'}`} size={25} /> </Nav.Link>
+    <MainSearchBar />
     <NavDropdown title='Account' active='true' className='metriq-navbar-text' alignRight>
       <NavDropdown.Item as={Link} to='/Profile' eventKey='15'><p className='font-weight-bold'>Settings</p></NavDropdown.Item>
       <NavDropdown.Item as={Link} to='/AddSubmission' eventKey='16'><p className='font-weight-bold'>Add Submission</p></NavDropdown.Item>
