+++ src/components/MainNavLeft.js
@@ -1,94 +1,9 @@
-import axios from 'axios'
-import React, { useEffect, useState } from 'react'
+import React from 'react'
 import { Nav, NavDropdown } from 'react-bootstrap'
-import { NavLink, useHistory } from 'react-router-dom'
-import FormFieldTypeaheadRow from '../components/FormFieldTypeaheadRow'
-import config from './../config'
+import { NavLink } from 'react-router-dom'
+import { FaDiscord, FaGithub, FaTwitter } from 'react-icons/fa'
 
 const MainNavLeft = () => {
-  const [allNames, setAllNames] = useState([])
-  const [taskNames, setTaskNames] = useState([])
-  const [methodNames, setMethodNames] = useState([])
-  const [platformNames, setPlatformNames] = useState([])
-  const [tagNames, setTagNames] = useState([])
-  const [submissionNames, setSubmissionNames] = useState([])
-  const history = useHistory()
-
-  useEffect(() => {
-    if (allNames.length > 0) {
-      return
-    }
-
-    axios.get(config.api.getUriPrefix() + '/task/names')
-      .then(res => {
-        const tNames = res.data.data
-        setTaskNames(tNames)
-
-        axios.get(config.api.getUriPrefix() + '/method/names')
-          .then(res => {
-            const mNames = res.data.data
-            setMethodNames(mNames)
-
-            axios.get(config.api.getUriPrefix() + '/platform/names')
-              .then(res => {
-                const pNames = res.data.data
-                setPlatformNames(pNames)
-
-                axios.get(config.api.getUriPrefix() + '/tag/names')
-                  .then(res => {
-                    const tgNames = res.data.data
-                    setTagNames(tgNames)
-
-                    axios.get(config.api.getUriPrefix() + '/submission/names')
-                      .then(res => {
-                        const sNames = res.data.data
-                        setSubmissionNames(sNames)
-
-                        setAllNames(tNames.concat(mNames).concat(pNames).concat(tgNames).concat(sNames))
-                      })
-                      .catch(err => {
-                        console.log(err)
-                      })
-                  })
-              })
-              .catch(err => {
-                console.log(err)
-              })
-          })
-          .catch(err => {
-            console.log(err)
-          })
-      })
-      .catch(err => {
-        console.log(err)
-      })
-  }, [allNames, setAllNames,
-    taskNames, setTaskNames,
-    methodNames, setMethodNames,
-    platformNames, setPlatformNames,
-    submissionNames, setSubmissionNames])
-
-  const handleOnSelect = (value) => {
-    if (!value) {
-      return
-    }
-    if (taskNames.includes(value)) {
-      history.push('/Task/' + value.id)
-    }
-    if (methodNames.includes(value)) {
-      history.push('/Method/' + value.id)
-    }
-    if (platformNames.includes(value)) {
-      history.push('/Platform/' + value.id)
-    }
-    if (tagNames.includes(value)) {
-      history.push('/Tag/' + encodeURIComponent(value.name))
-    }
-    if (submissionNames.includes(value)) {
-      history.push('/Submission/' + value.id)
-    }
-  }
-
   return (
     <Nav className='metriq-navbar'>
       <Nav.Link as={NavLink} activeClassName='active-navlink' to='/Tasks' className='metriq-navbar-text' eventKey='1'>Tasks</Nav.Link>
@@ -97,24 +12,16 @@ const MainNavLeft = () => {
       <Nav.Link as={NavLink} activeClassName='active-navlink' to='/Tags' className='metriq-navbar-text' eventKey='4'>Tags</Nav.Link>
       <Nav.Link as={NavLink} activeClassName='active-navlink' to='/Submissions' className='metriq-navbar-text' eventKey='5'>Submissions</Nav.Link>
       <Nav.Link as={NavLink} activeClassName='active-navlink' to='/QEDC' className='metriq-navbar-text' eventKey='10'>QED-C</Nav.Link>
-      <NavDropdown title='About' active='true' className='metriq-navbar-text' alignRight>
+      <NavDropdown id='metriq-navbar-dropdown' title='About' active='true' className='metriq-navbar-text' alignRight>
         <NavDropdown.Item as={NavLink} activeClassName='active-dropdown-navlink' to='/About' eventKey='6'><p className='font-weight-bold'>About</p></NavDropdown.Item>
         <NavDropdown.Item as={NavLink} activeClassName='active-dropdown-navlink' to='/Partners' eventKey='7'><p className='font-weight-bold'>Partners</p></NavDropdown.Item>
         <NavDropdown.Item as={NavLink} activeClassName='active-dropdown-navlink' to='/FAQ' eventKey='8'><p className='font-weight-bold'>F.A.Q.</p></NavDropdown.Item>
         <NavDropdown.Item as={NavLink} activeClassName='active-dropdown-navlink' to='/UserGuidelines' eventKey='9'><p className='font-weight-bold'>User Guidelines</p></NavDropdown.Item>
       </NavDropdown>
-      <div className='main-search-bar'>
-        <FormFieldTypeaheadRow
-          options={allNames}
-          labelKey='name'
-          inputName='name'
-          placeholder='Search...'
-          value=''
-          onSelect={handleOnSelect}
-          alignLabelRight
-          isWide
-        />
-      </div>
+      <span className='metriq-navbar-spacer' />
+      <Nav.Link href='https://twitter.com/MetriqInfo' eventKey='10'> <FaTwitter color={`${'white'}`} size={25} className='metriq-navbar-social-icon' /> </Nav.Link>
+      <Nav.Link href='http://discord.unitary.fund' eventKey='11'> <FaDiscord color={`${'white'}`} size={25} className='metriq-navbar-social-icon' /> </Nav.Link>
+      <Nav.Link href='https://github.com/unitaryfund/metriq-app' eventKey='12'> <FaGithub color={`${'white'}`} size={25} className='metriq-navbar-social-icon' /> </Nav.Link>
     </Nav>
   )
 }
